+++ browser-extension/src/legadilo.js
@@ -34,7 +34,16 @@ export const saveArticle = async ({ url, title, content }) => {
     throw new Error("Invalid url");
   }
 
-  return await post("/api/reading/articles/", { url, title, content });
+  try {
+    return await post("/api/reading/articles/", { url, title, content });
+  } catch (error) {
+    if (error.message === "Response status: 400 (Bad Request)") {
+      // Content might be too big (we don't have a clean way to catch this yet), let's try to save only the URL.
+      return await post("/api/reading/articles/", { url });
+    }
+
+    throw error;
+  }
 };
 
 export const updateArticle = async (
@@ -163,7 +172,7 @@ const doFetch = async (url, fetchOptions) => {
   const resp = await fetch(`${options.instanceUrl}${url}`, fetchOptions);
 
   if (!resp.ok) {
-    throw new Error(`Response status: ${resp.status} (${resp.statusText})`, { cause: resp.status });
+    throw new Error(`Response status: ${resp.status} (${resp.statusText})`);
   }
 
   if (fetchOptions.method === "DELETE") {